requirements.txt
dataclasses
datetime
google-cloud-storage
ipython
matplotlib
numpy
xarray
cartopy
dask